refactor(analysis-types): move fun serde test
